foodify/pyproject.toml
[tool.poetry]
name = "foodify"
version = "0.1.0"
description = ""
authors = ["Luis Galdeano <lpgaldeano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
sqlalchemy = "^2.0.12"
pandas = "^2.0.1"
openfoodfacts = "^0.1.3"
functions = "^0.7.0"
webdriver-manager = "^3.8.6"
selenium = "^4.9.0"
bs4 = "^0.0.1"
opencv-python = "^4.7.0.72"
certifi = "^2023.5.7"
numpy = "^1.24.3"
opencv-contrib-python = "^4.7.0.72"
pyzbar = "^0.1.9"
pydub = "^0.25.1"
ffmpeg = "^1.4"
uvicorn = "^0.22.0"
fastapi = "^0.103.0"
sqlalchemy-utils = "^0.41.1"
python-dotenv = "^1.0.0"
schedule = "^1.2.0"
apscheduler = "^3.10.1"
starlette = "^0.27.0"
fpdf = "^1.7.2"
psycopg2-binary = "^2.9.6"
pydantic = {extras = ["email"], version = "^2.6.0"}
autopep8 = "^2.0.4"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.270"
isort = "^5.12.0"
pylint = "^2.17.4"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
max-line-length = 90
disable = ["C0114", "W0718", "R0902", "R0903"]
good-names = ['ex', ]

[tool.black]
line-length = 90
exclude = '"htmlcov", "venv", ".pytest_cache", ".ruff_cache", "docs"'

[tool.isort]
skip = "__init__.py"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
relative_files = true

[tool.ruff]
line-length = 120
exclude = [
    ".pytest_cache",
    "venv",
]
[tool.ruff.per-file-ignores]
"__init__.py" = ["F401", "F403"]